+++ aws/driver/gen_template_defs.go
@@ -90,6 +90,8 @@ var APIPerTemplateDefName = map[string]string{
 	"checkdatabase":             "rds",
 	"createdbsubnetgroup":       "rds",
 	"deletedbsubnetgroup":       "rds",
+	"createregistry":            "ecr",
+	"deleteregistry":            "ecr",
 	"createuser":                "iam",
 	"deleteuser":                "iam",
 	"attachuser":                "iam",
@@ -621,6 +623,20 @@ var AWSTemplatesDefinitions = map[string]template.Definition{
 		RequiredParams: []string{"name"},
 		ExtraParams:    []string{},
 	},
+	"createregistry": {
+		Action:         "create",
+		Entity:         "registry",
+		Api:            "ecr",
+		RequiredParams: []string{"name"},
+		ExtraParams:    []string{},
+	},
+	"deleteregistry": {
+		Action:         "delete",
+		Entity:         "registry",
+		Api:            "ecr",
+		RequiredParams: []string{"name"},
+		ExtraParams:    []string{"account", "force"},
+	},
 	"createuser": {
 		Action:         "create",
 		Entity:         "user",
@@ -1057,6 +1073,8 @@ func DriverSupportedActions() map[string][]string {
 	supported["check"] = append(supported["check"], "database")
 	supported["create"] = append(supported["create"], "dbsubnetgroup")
 	supported["delete"] = append(supported["delete"], "dbsubnetgroup")
+	supported["create"] = append(supported["create"], "registry")
+	supported["delete"] = append(supported["delete"], "registry")
 	supported["create"] = append(supported["create"], "user")
 	supported["delete"] = append(supported["delete"], "user")
 	supported["attach"] = append(supported["attach"], "user")